Vital bug of the circular queue of the interpreter fixed
	The point of the bug is that the queue isn't really circular and the queue become invalid when the front counter reached max_size - 1.

 Changes to be committed:
	modified:   USER/interpreter.c
	modified:   USER/interpreter.h
	modified:   USER/utils.h

+++ USER/utils.h
@@ -14,7 +14,7 @@ double int2double(signed int i);
 typedef struct
 {
 	char *data;
-	uint16_t front, rear;
+	uint32_t front, rear;
 	uint16_t count, max_size;
 }	char_queue;
